+++ test/inject_test.dart
@@ -33,7 +33,19 @@ void main() {
   });
   group("inject tests", () {
     test("inject one-level", () {
-      ObjectUtil.inject(new User(), {
+      var user = ObjectUtil.inject(new User(), {
+        "firstName": "Bill",
+        "lastName": "Gates",
+        "age": "32" //test coercion
+      });
+
+      expect(user.firstName, "Bill");
+      expect(user.lastName, "Gates");
+      expect(user.age, 32);
+    });
+
+    test("inject one-level (async)", () {
+      ObjectUtil.injectAsync(new User(), {
         "firstName": "Bill",
         "lastName": "Gates",
         "age": "32" //test coercion
@@ -45,7 +57,26 @@ void main() {
     });
 
     test("inject three-level and auto-assign", () {
-      ObjectUtil.inject(new User(), {
+      var user = ObjectUtil.inject(new User(), {
+        "firstName": "Bill",
+        "lastName": "Gates",
+        "age": "32", //test coercion
+        "manager.firstName": "John",
+        "manager.lastName": "Kyle",
+        "manager.manager.firstName": "Boss"
+      });
+
+      expect(user.firstName, "Bill");
+      expect(user.lastName, "Gates");
+      expect(user.age, 32);
+      expect(user.manager, isNotNull);
+      expect(user.manager.firstName, "John");
+      expect(user.manager.lastName, "Kyle");
+      expect(user.manager.manager.firstName, "Boss");
+    });
+
+    test("inject three-level and auto-assign (async)", () {
+      ObjectUtil.injectAsync(new User(), {
         "firstName": "Bill",
         "lastName": "Gates",
         "age": "32", //test coercion
@@ -64,7 +95,20 @@ void main() {
     });
 
     test("inject non-existing", () {
-      ObjectUtil.inject(new User(), {
+      var user = ObjectUtil.inject(new User(), {
+        "firstName": "Bill",
+        "nonExisting": "Gates",
+        "level1.level2": 123,
+        "age": "32" //test coercion
+      }, silent: true);
+
+      expect(user.firstName, "Bill");
+      expect(user.lastName, isNull);
+      expect(user.age, 32);
+    });
+
+    test("inject non-existing (async)", () {
+      ObjectUtil.injectAsync(new User(), {
         "firstName": "Bill",
         "nonExisting": "Gates",
         "level1.level2": 123,
@@ -79,7 +123,29 @@ void main() {
     test("inject validate and onSetterError", () {
       List<String> validated = [];
       List<String> setterFailed = [];
-      ObjectUtil.inject(new User(), {
+      var user = ObjectUtil.inject(new User(), {
+        "firstName": "Bill",
+        "lastName": "Gates",
+        "age": "32", //test coercion
+        "notFound": false,
+        "wrong": true,
+      }, validate: (obj, field, value) {
+        validated.add(field);
+      }, onSetterError: (obj, field, value, error) {
+        setterFailed.add(field);
+      }, silent: true);
+
+      expect(user.firstName, "Bill");
+      expect(user.lastName, "Gates");
+      expect(user.age, 32);
+      expect(validated, ["firstName", "lastName", "age", "wrong"]);
+      expect(setterFailed, ["wrong"]);
+    });
+
+    test("inject validate and onSetterError (async)", () {
+      List<String> validated = [];
+      List<String> setterFailed = [];
+      ObjectUtil.injectAsync(new User(), {
         "firstName": "Bill",
         "lastName": "Gates",
         "age": "32", //test coercion
@@ -100,7 +166,24 @@ void main() {
 
     test("inject two-level and onSetterError", () {
       List<String> setterFailed = [];
-      ObjectUtil.inject(new User(), {
+      var user = ObjectUtil.inject(new User(), {
+        "manager.wrong": false,
+        "whatever": 123,
+        "manager.firstName": "John",
+        "manager.firstName.notFound": true,
+        "getterOnly.firstName": "Impossible",
+        "wrongUser.lastName": "Oops"
+      }, onSetterError: (obj, field, value, error) {
+        setterFailed.add(field);
+      });
+
+      expect(user.manager.firstName, "John");
+      expect(setterFailed, ["wrong", "whatever", "notFound", "getterOnly", "wrongUser"]);
+    });
+
+    test("inject two-level and onSetterError (async)", () {
+      List<String> setterFailed = [];
+      ObjectUtil.injectAsync(new User(), {
         "manager.wrong": false,
         "whatever": 123,
         "manager.firstName": "John",